Oracle estimator of the Logistic FAR problem

Logistic_FAR_Oracle computes the oracle estimator of the Logistic FAR problem.

Usage

Logistic_FAR_Oracle(
  y_vec,
  x_mat,
  h,
  k_n,
  p,
  mu_2,
  h_inv,
  eta_inv_stack,
  delta_init,
  eta_stack_init,
  mu_1_init,
  tol,
  max_iter
)

Arguments

y_vec

response vector, 0 for control, 1 for case. n = length(y_vec) is the number of observations.

x_mat

covariate matrix, consists of two parts. dim(x_mat) = (n, h + p * kn) First h columns are for demographical covariates. End user can include an intercept term. Rest columns are for p functional covariates, each being represented by a set of basis functions resulting kn covariates. Hence there are p * kn columns for the functional part.

h, kn, p

dimension information for the dataset(x_mat).

h_inv, eta_inv_stack

intermediate variables in the algorithm. If not provided, then it can be computed from the dataset.

delta_init, eta_stack_init, mu_1_init

initial values for the algorithm.

tol, max_iter

convergence tolerance and max number of iteration of the algorithm.

mu2

quadratic term in the ADMM algorithm

Details

In high-dimensional statistics, the oracle estimator is achieved by assuming the real model structure is known, i.e. an oracle knows what are the relevant covariates in the model. And the covariate effects are computed through unpenalized regression.

Basically, this functions performs unpenalized Logistic FAR (with log-contrast constrain) on the given dataset y_vec and x_mat. Hence it is end user who acts like an oracle to decide which variables will be passed to this function. Sometimes, this procedure is similar to a post-selection estimation.