Post-selection estimation

This function performs post-selection estimation on a given solution.

Usage

Logistic_FAR_Path_Further_Improve(
  x_mat,
  y_vec,
  h,
  k_n,
  p,
  delta_vec_init,
  eta_stack_init,
  mu1_vec_init,
  mu2,
  a = 1,
  lam = 0.1,
  weight_vec = 1,
  logit_weight_vec = 1,
  weight_already_combine = FALSE,
  tol = 10^(-5),
  max_iter = 1000,
  fast_glm = FALSE
)

Arguments

x_mat

covariate matrix, consists of two parts. dim(x_mat) = (n, h + p * kn) First h columns are for demographical covariates(can include an intercept term) Rest columns are for p functional covariates, each being represented by a set of basis functions resulting kn covariates.

y_vec

response vector, 0 for control, 1 for case. n = length(y_vec) is the number of observations.

h, k_n, p

dimension information for the dataset(x_mat).

delta_vec_init, eta_stack_init, mu1_vec_init

Initial values for the algorithm. This function uses these initial values to find out the active functional covariates. And the post-selection estimation begins with these initial values.

mu2

quadratic term in the ADMM algorithm

a

parameters for the algorithm. The 1st term in the loss function is 1 / a * loglik. See Algorithm_Details.pdf for more information.

lam

A scalar for the regularize in ridge penalty form in case of model saturation.

tol, max_iter

convergence tolerance and max number of iteration of the algorithm.